[solver] Swap smt_conv/smt_solver file names to match contents
After the slim/fat split the file names were inverted: smt_conv.{h,cpp}
held the fat smt_solver_baset while smt_solver.h held the slim smt_convt.
Realign so each file is named for what it defines: the fat
implementation moves to smt_solver.{h,cpp}, and smt_conv.h now holds the
slim smt_convt wrapper, including smt_solver.h for the complete base.

Pure rename/move; no behavioural change. Callers needing only the slim
interface include smt_conv.h; backends and solver internals include
smt_solver.h directly.
