+++ src/goto-symex/pytest.cpp
@@ -3,7 +3,7 @@
 #include <ac_config.h>
 #include <irep2/irep2.h>
 #include <irep2/irep2_expr.h>
-#include <solvers/smt/smt_solver.h>
+#include <solvers/smt/smt_conv.h>
 #include <util/c_types.h>
 #include <util/message.h>
 #include <util/config.h>
