+++ src/solvers/smt/CMakeLists.txt
@@ -6,7 +6,7 @@ add_library(smt
   array_conv.cpp
   smt_byteops.cpp
   smt_casts.cpp
-  smt_conv.cpp
+  smt_solver.cpp
   # IEEE754 constants/semantics and FP predicates extracted from smt_conv.cpp
   smt_fp_conv.cpp
   smt_memspace.cpp
