+++ src/solvers/smt/smt_solver.cpp
@@ -2,7 +2,7 @@
 #include <cfloat>
 #include <iomanip>
 #include <solvers/prop/literal.h>
-#include <solvers/smt/smt_conv.h>
+#include <solvers/smt/smt_solver.h>
 #include <solvers/smt/smt_fp_rounding_utils.h>
 #include <sstream>
 #include <util/arith_tools.h>
